Replaced mission system by a component
